requirements.txt
flask_sqlalchemy
flask_login
flask
requests
python-dotenv
werkzeug
psycopg2
APScheduler